requirements.txt
langchain
langchain[google-genai]
fastapi
uvicorn[standard]
python-dotenv
pydantic
chromadb        
pymongo
faiss-cpu
langgraph
streamlit
langchain-community
langchain-core 
pypdf
python-multipart
requests
pymupdf 
pdf2image 
pytesseract 
pillow
